requirements.txt
Django==1.9.5
snownlp==0.12.3
ImageHash==1.0
numpy==1.11.0
Pillow==3.2.0
pytesseract==0.1.6
jieba==0.38
